tests: Add integration test to create Tensorboard
Test that the Tensorboard Controller is deployed properly, by creating a
Tensorboard and making sure that it is ready. More specifically, we test
whether the Tensorboard resource exists and wait until is has 1 ready
replica. To create the dummy Tensorboard we use the resource templates
added under 'tests/assets'. The PVC is mandatory for the successful
deployment of the Tensorboard, holding the latter's logs.

Signed-off-by: Phoevos Kalemkeris <phoevos.kalemkeris@canonical.com>