ci, refactor: use data-platform-workflows in on_pull/push workflows (#159)
* ci, refactor: use data-platform-workflows in on_pull/push workflows

This commit enables the usage of build_charm.yaml workflow, which allows independent and cached charm
builds, making the CI more efficient and fast.
This commit also refactors the tests and the integrate workflow to be able to use the charm artefacts
generated by the build_charm.yaml workflow.
In summary this change:
1. Uses build_charm.yaml in on_{pull&push}.yaml to enable individual cached builds
2. Refactors individual and repository level integration tests to be able to use charm artefacts besides building them
3. Modifies integrate.yaml to download the charm artefacts and uses them in the tox integrate command

Fixes #155