fix: fetch istio_gateway_info lib with charmcraft to match published lib (#166) (#168)
This repository contained a version of the istio_gateway_info lib that was not
exactly the one that was published in Charmhub. This caused issues when trying
to fetch new versions of the library.

This commit replaces the offending file with the library that is actually published.

Part of canonical/bundle-kubeflow#1209